feat: normalize grid layout
