+++ app/home/page.tsx
@@ -33,6 +33,7 @@ import {
   maxW,
   minH,
   maxH,
+  normalizeLayout,
 } from '@/lib/layout';
 
 const ReactGridLayout = WidthProvider(GridLayout);
@@ -73,57 +74,6 @@ const sortTplFields = (tpl: Template): Template => ({
   fields: tpl.fields.slice().sort((a, b) => a.y - b.y),
 });
 
-type LayoutItem = Layout;
-
-/**
- * Normalize layout items by sorting them visually (top-to-bottom, left-to-right)
- * and reassigning `x`/`y` coordinates to remove gaps and overlaps.
- *
- * This ensures a deterministic layout order regardless of how items were moved
- * around during editing. The algorithm greedily places each item in the first
- * available spot scanning rows left-to-right.
- */
-const normalizeLayout = (items: LayoutItem[], cols: number): LayoutItem[] => {
-  const sorted = items.slice().sort((a, b) => (a.y - b.y) || (a.x - b.x));
-  const occupied: boolean[][] = [];
-
-  const canPlace = (x: number, y: number, w: number, h: number) => {
-    for (let yy = y; yy < y + h; yy++) {
-      for (let xx = x; xx < x + w; xx++) {
-        if (occupied[yy]?.[xx]) return false;
-      }
-    }
-    return true;
-  };
-
-  const occupy = (x: number, y: number, w: number, h: number) => {
-    for (let yy = y; yy < y + h; yy++) {
-      if (!occupied[yy]) occupied[yy] = [];
-      for (let xx = x; xx < x + w; xx++) {
-        occupied[yy][xx] = true;
-      }
-    }
-  };
-
-  return sorted.map((item) => {
-    const w = item.w ?? 1;
-    const h = item.h ?? 1;
-    let x = 0;
-    let y = 0;
-    while (true) {
-      if (x + w > cols) {
-        x = 0;
-        y += 1;
-        continue;
-      }
-      if (canPlace(x, y, w, h)) {
-        occupy(x, y, w, h);
-        return { ...item, x, y };
-      }
-      x += 1;
-    }
-  });
-};
 
 const Badge = ({ children }: { children: React.ReactNode }) => (
   <span className="px-2 py-0.5 rounded-full text-xs bg-slate-800 text-white/90">{children}</span>
